requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
django-rest-auth==0.9.5
django-allauth==0.44.0